requirements/base.txt
Django>=3.2,<4.2
djangorestframework
gunicorn
python-decouple
djangorestframework-simplejwt
